third_party/WebKit/LayoutTests/fast/files/url-null-expected.txt

Test URL methods with null and undefined arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS URL.createObjectURL(null) threw exception TypeError: Failed to execute 'createObjectURL' on 'URL': No function was found that matched the signature provided..
PASS URL.createObjectURL(undefined) threw exception TypeError: Failed to execute 'createObjectURL' on 'URL': No function was found that matched the signature provided..
PASS URL.revokeObjectURL(null) did not throw exception.
PASS URL.revokeObjectURL(undefined) did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
